requirements.txt
flask
metta-py
flask
pycardano
blockfrost-python
python-dotenv
